requirements.txt
Babel==0.9.6
Django==1.5.1
Mezzanine==1.4.10
Pillow==1.7.8
South==0.7.6
Unidecode==0.04.13
bleach==1.2.2
django-appconf==0.6
django-compressor==1.3
django-extensions==1.1.1
django-extra-views==0.5.5
django-haystack==2.0.0
django-oscar==0.5.1
django-smtp-ssl==1.0
django-treebeard==1.61
feedparser==5.1.3
filebrowser-safe==0.2.28
grappelli-safe==0.2.22
html5lib==0.95
ipdb==0.7
ipython==0.13.2
oauthlib==0.5.1
python-memcached==1.48
pytz==2013b
readline==6.2.4.1
requests==1.2.3
requests-oauthlib==0.3.2
six==1.3.0
sorl-thumbnail==11.12
wsgiref==0.1.2
pyOpenSSL==0.13
PyCrypto==2.6
djangorestframework